+++ app/api/__init__.py
@@ -1,7 +1,8 @@
 from flask_restplus import Api
+from app.users.views import API as users_api
+from app.posts.views import API as posts_api
 
-API = Api(title='Board API',
-        version='1.0',
-        description="Board REST API"
-        # All API metadatas
-    )
+REST_API = Api()
+
+REST_API.add_namespace(users_api, '/users')
+REST_API.add_namespace(posts_api, '/posts')
